requirements.txt
ultralytics
transformers
Pillow
easyocr
streamlit
opencv-python
